+++ src/main/java/blusunrize/immersiveengineering/common/blocks/metal/conveyors/ExtractConveyor.java
@@ -15,7 +15,6 @@ import blusunrize.immersiveengineering.api.tool.ConveyorHandler.ConveyorDirectio
 import blusunrize.immersiveengineering.client.ClientUtils;
 import blusunrize.immersiveengineering.common.util.Utils;
 import blusunrize.immersiveengineering.common.util.chickenbones.Matrix4;
-import com.google.common.collect.Lists;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.client.renderer.model.BakedQuad;
@@ -321,6 +320,7 @@ public class ExtractConveyor extends BasicConveyor
 		CompoundNBT nbt = super.writeConveyorNBT();
 		nbt.putInt("transferCooldown", transferCooldown);
 		nbt.putInt("transferTickrate", transferTickrate);
+		nbt.putInt("relativeExtractDir", relativeExtractDir.ordinal());
 		return nbt;
 	}
 
@@ -330,6 +330,9 @@ public class ExtractConveyor extends BasicConveyor
 		super.readConveyorNBT(nbt);
 		transferCooldown = nbt.getInt("transferCooldown");
 		transferTickrate = nbt.getInt("transferTickrate");
+		relativeExtractDir = Rotation.values()[nbt.getInt("relativeExtractDir")];
+		if(relativeExtractDir==Rotation.NONE)
+			relativeExtractDir = Rotation.CLOCKWISE_180;
 	}
 
 	private Direction getExtractDirection()